requirements.txt
aiohttp==3.11.12
Brotli==1.1.0
cachetools==5.5.1
chardet==5.1.0
Cython==3.0.12
google_api_python_client==2.167.0
google_auth_httplib2==0.2.0
h2==4.2.0
httplib2==0.22.0
imgkit==1.2.3
importlib_metadata==8.6.1
ipython==8.32.0
itsdangerous==2.2.0
networkx==3.4.2
oauthlib==3.2.2
peewee==3.17.9
pulp==3.1.1
PyJWT==2.10.1
pyppeteer==2.0.0
pytest==8.4.2
PyYAML==6.0.2
quart==0.20.0
quart_auth==0.11.0
quart_schema==0.22.0
requests_oauthlib==2.0.0
thefuzz==0.22.1
toml==0.10.2
tqdm==4.67.1
Werkzeug==3.1.3
wsproto==1.2.0
asyncpg==0.30.0
pytest~=8.4.2
pytest-asyncio~=1.2.0
redis~=7.0.1
sqlalchemy~=2.0.40
neo4j~=6.0.2